Refactor FFI examples to consolidate several examples into one submodule.
