+++ examples/ffi/src/jax_ffi_example/cpu_examples.cc
@@ -14,12 +14,26 @@ limitations under the License.
 ==============================================================================*/
 
 #include <cstdint>
+#include <mutex>
+#include <string_view>
+#include <unordered_map>
 
 #include "nanobind/nanobind.h"
 #include "xla/ffi/api/ffi.h"
 
 namespace nb = nanobind;
 namespace ffi = xla::ffi;
+
+// ----------
+// Attributes
+// ----------
+//
+// An example demonstrating the different ways that attributes can be passed to
+// the FFI.
+//
+// For example, we can pass arrays, variadic attributes, and user-defined types.
+// Full support of user-defined types isn't yet supported by XLA, so that
+// example will be added in the future.
 
 ffi::Error ArrayAttrImpl(ffi::Span<const int32_t> array,
                          ffi::ResultBufferR0<ffi::S32> res) {
@@ -54,13 +68,52 @@ XLA_FFI_DEFINE_HANDLER_SYMBOL(DictionaryAttr, DictionaryAttrImpl,
                                   .Ret<ffi::BufferR0<ffi::S32>>()
                                   .Ret<ffi::BufferR0<ffi::S32>>());
 
-NB_MODULE(_attrs, m) {
+// -------
+// Counter
+// -------
+//
+// An example demonstrating how an FFI call can maintain "state" between calls
+//
+// In this case, the ``Counter`` call simply accumulates the number of times it
+// was executed, but this pattern can also be used for more advanced use cases.
+// For example, this pattern is used in jaxlib for:
+//
+// 1. The GPU solver linear algebra kernels which require an expensive "handler"
+//    initialization, and
+// 2. The ``triton_call`` function which caches the compiled triton modules
+//    after their first use.
+
+ffi::Error CounterImpl(int64_t index, ffi::ResultBufferR0<ffi::S32> out) {
+  static std::mutex mutex;
+  static auto &cache = *new std::unordered_map<int64_t, int32_t>();
+  {
+    const std::lock_guard<std::mutex> lock(mutex);
+    auto it = cache.find(index);
+    if (it != cache.end()) {
+      out->typed_data()[0] = ++it->second;
+    } else {
+      cache.insert({index, 0});
+      out->typed_data()[0] = 0;
+    }
+  }
+  return ffi::Error::Success();
+}
+
+XLA_FFI_DEFINE_HANDLER_SYMBOL(
+    Counter, CounterImpl,
+    ffi::Ffi::Bind().Attr<int64_t>("index").Ret<ffi::BufferR0<ffi::S32>>());
+
+// Boilerplate for exposing handlers to Python
+NB_MODULE(_cpu_examples, m) {
   m.def("registrations", []() {
     nb::dict registrations;
     registrations["array_attr"] =
         nb::capsule(reinterpret_cast<void *>(ArrayAttr));
     registrations["dictionary_attr"] =
         nb::capsule(reinterpret_cast<void *>(DictionaryAttr));
+
+    registrations["counter"] = nb::capsule(reinterpret_cast<void *>(Counter));
+
     return registrations;
   });
 }
