+++ examples/ffi/tests/cpu_examples_test.py
@@ -18,7 +18,7 @@ import jax
 import jax.numpy as jnp
 from jax._src import test_util as jtu
 
-from jax_ffi_example import attrs
+from jax_ffi_example import cpu_examples
 
 jax.config.parse_flags_with_absl()
 
@@ -30,11 +30,11 @@ class AttrsTests(jtu.JaxTestCase):
       self.skipTest("Unsupported platform")
 
   def test_array_attr(self):
-    self.assertEqual(attrs.array_attr(5), jnp.arange(5).sum())
-    self.assertEqual(attrs.array_attr(3), jnp.arange(3).sum())
+    self.assertEqual(cpu_examples.array_attr(5), jnp.arange(5).sum())
+    self.assertEqual(cpu_examples.array_attr(3), jnp.arange(3).sum())
 
   def test_array_attr_jit_cache(self):
-    jit_array_attr = jax.jit(attrs.array_attr, static_argnums=(0,))
+    jit_array_attr = jax.jit(cpu_examples.array_attr, static_argnums=(0,))
     with jtu.count_jit_and_pmap_lowerings() as count:
       jit_array_attr(5)
     self.assertEqual(count[0], 1)  # compiles once the first time
@@ -44,22 +44,51 @@ class AttrsTests(jtu.JaxTestCase):
 
   def test_array_attr_no_jit(self):
     with jax.disable_jit():
-      attrs.array_attr(5)  # doesn't crash
+      cpu_examples.array_attr(5)  # doesn't crash
 
   def test_dictionary_attr(self):
-    secret, count = attrs.dictionary_attr(secret=5)
+    secret, count = cpu_examples.dictionary_attr(secret=5)
     self.assertEqual(secret, 5)
     self.assertEqual(count, 1)
 
-    secret, count = attrs.dictionary_attr(secret=3, a_string="hello")
+    secret, count = cpu_examples.dictionary_attr(secret=3, a_string="hello")
     self.assertEqual(secret, 3)
     self.assertEqual(count, 2)
 
     with self.assertRaisesRegex(Exception, "Unexpected attribute"):
-      attrs.dictionary_attr()
+      cpu_examples.dictionary_attr()
 
     with self.assertRaisesRegex(Exception, "Wrong attribute type"):
-      attrs.dictionary_attr(secret="invalid")
+      cpu_examples.dictionary_attr(secret="invalid")
+
+
+class CounterTests(jtu.JaxTestCase):
+  def setUp(self):
+    super().setUp()
+    if not jtu.test_device_matches(["cpu"]):
+      self.skipTest("Unsupported platform")
+
+  def test_basic(self):
+    self.assertEqual(cpu_examples.counter(0), 0)
+    self.assertEqual(cpu_examples.counter(0), 1)
+    self.assertEqual(cpu_examples.counter(0), 2)
+    self.assertEqual(cpu_examples.counter(1), 0)
+    self.assertEqual(cpu_examples.counter(0), 3)
+
+  def test_jit(self):
+    @jax.jit
+    def counter_fun(x):
+      return x, cpu_examples.counter(2)
+
+    self.assertEqual(counter_fun(0)[1], 0)
+    self.assertEqual(counter_fun(0)[1], 1)
+
+    # Persists across different cache hits
+    self.assertEqual(counter_fun(1)[1], 2)
+
+    # Persists after the cache is cleared
+    counter_fun.clear_cache()
+    self.assertEqual(counter_fun(0)[1], 3)
 
 
 if __name__ == "__main__":
